analyser/requirements.txt
langchain==0.1.0
langchain-openai==0.1.17
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2

# OpenAI
openai==1.3.7

# Async support
aiohttp==3.9.1

# Environment and configuration
python-dotenv==1.0.0

# Vector database (optional for testing)
chromadb==0.4.22

# Basic data processing (without pandas for now)
numpy==1.25.2

# PDF processing
PyPDF2==3.0.1 